Second

+++ WebMauGiao/WebMauGiao.Service/Infrastructure/IRepository.cs
@@ -21,7 +21,7 @@ namespace WebMauGiao.Service.Infrastructure
         void DeleteMulti(Expression<Func<T, bool>> where);
 
         // Get an entity by int id
-        T GetSingleById(int id);
+        T GetSingleById(long id);
 
         T GetSingleByCondition(Expression<Func<T, bool>> expression, string[] includes = null);
